Make the dashboard view self-contained.
Add a requirement to test edge case of nil acquirement on the dashboard.
A table cannot use a partial render (it is not as happy) so it is better
to put all the table code in the same file. It also fix the duplication
of the table in the output.
Now the test page shows two missing skills so we need to specify better
the row we use for testing.